Minor cosmetic tweaks

+++ fbink_device_id.h
@@ -72,7 +72,7 @@ typedef struct __attribute__((__packed__))
 static bool     is_kindle_device(uint32_t);
 static bool     is_kindle_device_new(uint32_t);
 static uint32_t from_base(const char*, uint8_t);
-static char*    to_base(int64_t, unsigned int);
+static char*    to_base(int64_t, uint8_t);
 static void     identify_kindle(void);
 #        elif defined(FBINK_FOR_CERVANTES)
 static void identify_cervantes(void);
